Avoid circular import Template -> Hooks -> Renderer -> Template
This commit does not actually fix the circular dependency in the design. We just
restrict the offending import to type-checking, where it does not cause errors.
While this is not ideal design-wise, it allows us to have a `Hooks` object in
`Template` instead of a mere `Path` to the hooks directory.