💥 [services] Use cutty.json to determine the base revision for updates (#558)
* ✨ [projects] Store template revision in cutty.json

* 🚨 [functional] Mark test for `cutty update --skip` as XFAIL

* 🔨 [projects] Remove default argument for `ProjectConfig.revision`

* 🎨 [projects] Reformat function `ProjectRepository.link`

* 🔨 [projects] Slide `latest` branch creation in `ProjectRepository.link`

Slide the creation of the `cutty/latest` branch before the commit copying
cutty.json onto the target branch. This change has no observable change in
behavior except for rare circumstances where the final commit fails.

The rationale for this refactoring is that we need to extract the creation of
both the `cutty/latest` and `cutty/update` branches, for use in `update`. The
extracted region removes both branches initially, so `update` would fail without
the statement that (re-)creates the latest branch.

* 🔨 [projects] Split variable `update` in `ProjectRepository.link`

* 🔨 [projects] Extract function `ProjectRepository.reset`

* 🔨 [services] Slide statement in `update`

* ✨ [services] Use template revision from cutty.json in `cutty update`

* ✅ [functional] Remove XFAIL marker for reverted update test

* ✅ [functional] Remove XFAIL marker for `cutty update --skip` test

* 🔨 [projects] Extract variable `message` in `ProjectRepository.link`

* 🔨 [projects] Extract function `ProjectRepository.updateconfig`

* ✨ [projects] Update project config on `cutty update --skip`

* ✅ [projects] Adapt unit test for `ProjectRepository.skipupdate` to provide cutty.json