🔨 [functional] Simplify test for `cutty create` with existing files
The previous version was correct because pygit2 converts the Path to a str
before inspecting the tree, see `Tree_contains` in `src/tree.c`. Use a str for
consistency with similar tests and to avoid the impression that the version with
Path did not constitute a failing test (which it did).