💥 [entrypoints] Remove options `cutty create --skip-if-file-exists` and `--overwrite-if-exists` (#600)
* ✅ [functional] Add test for `cutty create` with untracked project files

* ✅ [functional] Add test for `cutty create` with existing project files

* ✅ [functional] Add test for `cutty create` with conflicting changes

* 💥 [entrypoints] Remove option `cutty create --overwrite-if-exists`

This option was broken in #585 when `cutty create` was switched to use
cherry-picking to introduce changes. As project creation currently always stores
the project in a git repository, it appears more practical to rely on git to
manage existing files.

We might eventually reintroduce this functionality if we ever add pure
disk-based storage for `cutty create`. In that case, we'd likely switch to a
different CLI, such as `--file-exists={error,keep,overwrite}`.

* 💥 [entrypoints] Remove option `cutty create --skip-if-file-exists`

* 🔨 [services] Remove parameter `fileexists` from `createproject`

* 🔨 [services] Rename functions `{create,cookiecutter}.create{project => }`