🔨 [packages] Clone Mercurial repositories with working directory
Without a working directory, we would need to specify an explicit revision for
`hg archive` when mounting the template package. This revision should be the
same revision that would be checked out by `hg clone`.

Unfortunately, determining this revision is not trivial. (IIUC it's the revision
pointed to by a bookmark named `@` if it exists, at least for modern clients,
and the tip-most revision of the `default` branch, otherwise.)

Avoid the complexity by letting Mercurial clone with a working directory. Then
`hg archive` can default to the working directory parent.