🔧 [coverage] Exclude some `pass` statements in tests from coverage
All of these can be safely excluded because the tests are designed to trigger an
exception before the `pass` is executed.

These cases of missing statement coverage are only caught since Python 3.10,
because a final `pass` was traced incorrectly on Python 3.9 and earlier. See
https://bugs.python.org/issue44672 for details, closed as "won't fix".

Note also that these cases are only caught when producing coverage data with
Python 3.10 *exclusively*. That's why we did not see them in CI even after
adding Python 3.10 support.